# scripts/scenarios/core/agent-metareasoning-manager/api/openapi.yaml
openapi: 3.0.3
info:
  title: Metareasoning API
  description: |
    A comprehensive API for managing and executing AI-powered metareasoning workflows.
    
    ## Features
    - **Workflow Management**: Create, update, and manage workflows across multiple platforms
    - **AI Integration**: Generate workflows using large language models
    - **Multi-Platform Support**: Execute workflows on n8n, Windmill, and other platforms
    - **Performance Optimized**: <100ms response times for most endpoints
    - **Import/Export**: Seamless workflow portability
    - **Search & Discovery**: Find workflows with semantic search
    
    ## Authentication
    All endpoints (except `/health`) require Bearer token authentication.
    
    ## Performance
    - Response times are tracked via `X-Response-Time` headers
    - Caching is indicated via `X-Cache` headers
    - Most endpoints target <100ms response time
  version: 3.0.0
  contact:
    name: Metareasoning API Support
    url: https://github.com/vrooli/vrooli
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8093
    description: Development server
  - url: https://api.metareasoning.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /workflows:
    get:
      summary: List Workflows
      description: Get paginated list of workflows with filtering
      tags:
        - Workflows
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [n8n, windmill, both]
          description: Filter by execution platform
        - name: type
          in: query
          schema:
            type: string
          description: Filter by workflow type
        - name: active
          in: query
          schema:
            type: boolean
            default: true
          description: Show only active workflows
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: List of workflows
          headers:
            X-Cache:
              schema:
                type: string
              description: Cache status (HIT/MISS)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'

    post:
      summary: Create Workflow
      description: Create a new workflow
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique workflow identifier
        name:
          type: string
          description: Human-readable workflow name
        description:
          type: string
          description: Detailed workflow description
        type:
          type: string
          description: Workflow category/type
        platform:
          type: string
          enum: [n8n, windmill, both]
          description: Target execution platform
        config:
          type: object
          description: Platform-specific configuration
        webhook_path:
          type: string
          nullable: true
          description: Webhook endpoint path
        job_path:
          type: string
          nullable: true
          description: Job execution path
        schema:
          type: object
          description: Input/output schema
        estimated_duration:
          type: integer
          description: Estimated execution time in ms
        version:
          type: integer
          description: Workflow version number
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: Parent workflow for versioning
        is_active:
          type: boolean
          description: Whether workflow is active
        is_builtin:
          type: boolean
          description: Whether workflow is built-in
        tags:
          type: array
          items:
            type: string
          description: Workflow tags for categorization
        usage_count:
          type: integer
          description: Number of times executed
        success_count:
          type: integer
          description: Number of successful executions
        failure_count:
          type: integer
          description: Number of failed executions
        avg_execution_time_ms:
          type: integer
          nullable: true
          description: Average execution time
        created_by:
          type: string
          description: Creator username
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    WorkflowCreate:
      type: object
      required:
        - name
        - type
        - platform
        - config
      properties:
        name:
          type: string
          maxLength: 100
          description: Workflow name
        description:
          type: string
          description: Workflow description
        type:
          type: string
          maxLength: 50
          description: Workflow type/category
        platform:
          type: string
          enum: [n8n, windmill, both]
          description: Target platform
        config:
          type: object
          description: Platform configuration
        webhook_path:
          type: string
          nullable: true
          description: Webhook path
        job_path:
          type: string
          nullable: true
          description: Job path
        schema:
          type: object
          description: Input/output schema
        estimated_duration:
          type: integer
          minimum: 0
          description: Estimated duration in ms
        tags:
          type: array
          items:
            type: string
          description: Workflow tags
...
    ListResponse:
      type: object
      properties:
        workflows:
          type: array
          items:
            $ref: '#/components/schemas/Workflow'
        total:
          type: integer
          description: Total number of workflows
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Items per page
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: Error code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...